+++ src/keygen/mod.rs
@@ -1,5 +1,5 @@
 pub mod password;
 pub mod phrase;
 
-pub use password::generate_password;
+pub use password::{generate_compatible_password, generate_password};
 pub use phrase::generate_phrase;